Move serialization types to `ser.rs`
.. a larger refactor, which however means LSPS0 module structure is
equivalent to others, and serialization-specific things are separated.
