+++ src/lsps1/service.rs
@@ -18,12 +18,11 @@ use super::msgs::{
 use super::utils::is_valid;
 use crate::message_queue::MessageQueue;
 
-use crate::events::EventQueue;
-use crate::lsps0::msgs::{ProtocolMessageHandler, RequestId};
+use crate::events::{Event, EventQueue};
+use crate::lsps0::ser::{ProtocolMessageHandler, RequestId, ResponseError};
 use crate::prelude::{HashMap, String, ToString};
 use crate::sync::{Arc, Mutex, RwLock};
 use crate::utils;
-use crate::{events::Event, lsps0::msgs::ResponseError};
 
 use lightning::chain::Filter;
 use lightning::ln::channelmanager::AChannelManager;
