+++ src/lsps2/msgs.rs
@@ -8,8 +8,9 @@ use bitcoin::hashes::{Hash, HashEngine};
 use chrono::Utc;
 use serde::{Deserialize, Serialize};
 
-use crate::lsps0::msgs::{LSPSMessage, RequestId, ResponseError};
-use crate::lsps0::ser::{string_amount, string_amount_option};
+use crate::lsps0::ser::{
+	string_amount, string_amount_option, LSPSMessage, RequestId, ResponseError,
+};
 use crate::prelude::{String, Vec};
 use crate::utils;
