+++ src/message_queue.rs
@@ -1,6 +1,6 @@
 //! Holds types and traits used to implement message queues for [`LSPSMessage`]s.
 
-use crate::lsps0::msgs::LSPSMessage;
+use crate::lsps0::ser::LSPSMessage;
 use crate::prelude::{Box, Vec, VecDeque};
 use crate::sync::{Mutex, RwLock};
